Place gui text at whole pixels
The text helpers take integer coordinates again, so the float casts around
the centred titles come off.

Error sites 1067 -> 1062.

+++ src/main/java/com/simibubi/create/content/equipment/clipboard/ClipboardScreen.java
@@ -547,7 +547,7 @@ public class ClipboardScreen extends AbstractSimiScreen {
 		if (!pIsEndOfText) {
 			graphics.fill(pCursorPos.x, pCursorPos.y - 1, pCursorPos.x + 1, pCursorPos.y + 9, -16777216);
 		} else {
-			graphics.text(font, "_", (float) pCursorPos.x, (float) pCursorPos.y, 0, false);
+			graphics.text(font, "_", pCursorPos.x, pCursorPos.y, 0, false);
 		}
 	}
